Revert "all network code in a manager that can be mocked according to an interface"
This reverts commit 0f1c71c7f864723901cf35ebae69ebc9f1d7c6c0.
